address some of  sonarQube issues

+++ src/util/dict/dict.c
@@ -171,7 +171,7 @@ dictType dictTypeHeapRedisStrings = {
 // Hash function for uint64_t keys stored directly as void* (cast).
 // Identity hash — sufficient for sequential integer keys.
 static uint64_t uint64HashFunction(const void *key) {
-  return (uint64_t)(uintptr_t)key;
+  return (uintptr_t)key;
 }
 
 // Dict type for uint64_t keys stored as void* (no dup/free needed).
